+++ src/taiko/rim.rs
@@ -1,14 +1,7 @@
-use crate::parse::{HitObject, HitSound};
+use crate::parse::HitSound;
 
 pub(crate) trait Rim {
     fn is_rim(&self) -> bool;
-}
-
-impl Rim for HitObject {
-    #[inline]
-    fn is_rim(&self) -> bool {
-        self.sound.clap() || self.sound.whistle()
-    }
 }
 
 impl Rim for u8 {