+++ src/taiko/stamina_cheese.rs
@@ -1,5 +1,5 @@
 use super::{LimitedQueue, Rim};
-use crate::{parse::HitObject, Beatmap};
+use crate::Beatmap;
 
 const ROLL_MIN_REPETITIONS: usize = 12;
 const TL_MIN_REPETITIONS: isize = 16;
@@ -31,7 +31,7 @@ impl StaminaCheeseDetector for Beatmap {
         let mut index_before_last_repeat = -1;
         let mut last_mark_end = 0;
 
-        for (i, h) in self.hit_objects.iter().enumerate() {
+        for (i, &h) in self.sounds.iter().enumerate() {
             history.push(h);
 
             if !history.full() {
@@ -62,8 +62,8 @@ impl StaminaCheeseDetector for Beatmap {
         let mut tl_len = -2;
         let mut last_mark_end = 0;
 
-        for (i, h) in self.hit_objects.iter().enumerate().skip(parity).step_by(2) {
-            if h.is_rim() == is_rin {
+        for (i, &sound) in self.sounds.iter().enumerate().skip(parity).step_by(2) {
+            if sound.is_rim() == is_rin {
                 tl_len += 2;
             } else {
                 tl_len = -2;
@@ -94,7 +94,7 @@ fn mark_as_cheese(start: usize, end: usize, cheese: &mut [bool]) {
 }
 
 #[inline]
-fn contains_pattern_repeat(history: &LimitedQueue<&HitObject>, pattern_len: usize) -> bool {
+fn contains_pattern_repeat(history: &LimitedQueue<u8>, pattern_len: usize) -> bool {
     for (&curr, &to_compare) in history.iter().zip(history.iter().skip(pattern_len)) {
         if curr.is_rim() != to_compare.is_rim() {
             return false;